fix(ci): make macOS-only probe examples compile cross-platform; format RESULTS.md
The crate-level cfg stripped the probes to empty files on Linux CI (E0601 no main). Real impls move to examples/probe_support/ (not auto-compiled) and are include!'d on macOS with a stub main elsewhere. RESULTS.md was prettier-noncompliant.
